Finished functionality for adding buildings to database
